LayoutTests/fast/dom/MutationObserver/added-out-of-order-expected.txt

Test MutationEvents interfering with MutationObservers: adding nodes 'out of order'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mutations.length is 2
PASS mutations[0].addedNodes.length is 1
PASS mutations[0].removedNodes.length is 0
PASS mutations[0].addedNodes[0].tagName is "DIV"
PASS mutations[1].addedNodes.length is 1
PASS mutations[1].removedNodes.length is 2
PASS mutations[1].addedNodes[0].nodeValue is "hello world"
PASS mutations[1].removedNodes[0].tagName is "SPAN"
PASS mutations[1].removedNodes[1].tagName is "DIV"
PASS successfullyParsed is true

TEST COMPLETE
